Removes json payload for resize.
Various renaming/cleanup/etc
