--- restricted_env_keys.go
@@ -1,6 +1,0 @@
-package main
-
-// A map of keys not to copy to the shell users env
-var restrictedEnvVars = map[string]bool{
-	"AUDIT_UPLOAD_URL": true,
-}
